requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Faker==24.4.0
Flask==3.0.2
Flask-Login==0.6.3
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.1
